Enable proper restart for xmonad
The change to the xmonad config forces it to always recompile on
restart, which seems to be necessary because a symlink into a nix store
file doesn't allow xmonad to detect when the config has changed.

The non-recompilation under the old config is the reason why the various
packages the xmonad config uses (alacritty, rofi, et al.) weren't
launching when the packages were removed from the system
configuration--the old xmonad config (that wasn't being replaced on
restart) refers to these packages just by the binary name, not the full
path.